+++ Editor/UI/BitmapFontCreatorEditor.cs
@@ -5,25 +5,26 @@ namespace dev.klebersilva.tools.bitmapfontcreator
 {
 	public class BitmapFontCreatorEditor : EditorWindow
 	{
-		public const string MenuItemPath = "Window/Bitmap Font Creator";
+		public const string MenuItemPath = "Window/Text/Bitmap Font Creator";
 
 		private readonly ExecutionData _data = ExecutionData.Default;
 		private CharacterPropsList _customCharPropsList;
 		private ProfilesView _profilesView;
 		private PrefsView _prefsView;
 		private Settings _settings;
+		private PrefsModel _prefs;
 		private string _error;
 
 		private Vector2 _charactersScrollPos = Vector2.zero;
 		private Vector2 _mainScrollPos = Vector2.zero;
 		private int _selectedCharacterSetIndex = 0;
 
-		private Vector2Int _guessCache;
+		private Vector2Int _guessRowsColsCache;
 
 		[MenuItem(MenuItemPath)]
 		public static void ShowWindow()
 		{
-			var size = new Vector2(300, 580);
+			var size = new Vector2(310, 640);
 			var window = GetWindowWithRect<BitmapFontCreatorEditor>(
 				new Rect((Screen.width - size.x) * 0.5f, (Screen.height - size.y) * 0.5f, size.x, size.y),
 				false,
@@ -38,37 +39,48 @@ namespace dev.klebersilva.tools.bitmapfontcreator
 		private void Setup()
 		{
 			_settings = Settings.Load();
-			var prefs = PrefsModel.Load();
+			_prefs = PrefsModel.Load();
 			_settings.Profiles.Selected?.CopyTo(_data);
 
 			_customCharPropsList = new CharacterPropsList(_data.CustomCharacterProps);
-			_profilesView = new ProfilesView(_data, _settings.Profiles, prefs);
-			_prefsView = new PrefsView(prefs);
+			_profilesView = new ProfilesView(_data, _settings.Profiles, _prefs);
+			_prefsView = new PrefsView(_prefs);
 		}
 
 		private void OnGUI()
 		{
-#if BITMAP_FONT_CREATOR_DEBUG
-			// Dev only
+#if BITMAP_FONT_CREATOR_DEV
 			if (_customCharPropsList == null) Setup();
 #endif
+
+			EditorGUIUtility.labelWidth = 120;
 			_mainScrollPos = GUILayout.BeginScrollView(_mainScrollPos, false, false, GUIStyle.none, GUI.skin.verticalScrollbar, GUILayout.ExpandHeight(true));
 			GUILayout.BeginVertical();
 
 			DrawTextureField();
 
-			_data.Cols = EditorGUILayout.IntField(UI.Cols, _data.Cols);
-			_data.Rows = EditorGUILayout.IntField(UI.Rows, _data.Rows);
+			_data.Cols = UIUtils.IntFieldMin(UI.Cols, _data.Cols, 1);
+			_data.Rows = UIUtils.IntFieldMin(UI.Rows, _data.Rows, 1);
 
 			GUILayout.BeginHorizontal();
-			GUILayout.Space(EditorGUIUtility.labelWidth);
-			if (GUILayout.Button(UI.GuessButton)) GuessRowsAndCols();
+			GUILayout.FlexibleSpace();
+			if (GUILayout.Button(UI.GuessRowsAndColsButton, Styles.MiniButton)) GuessRowsAndCols();
+			var showPreview = DrawPreviewButton();
 			GUILayout.EndHorizontal();
 
 			EditorGUILayout.Space();
 			_data.Orientation = (Orientation)EditorGUILayout.EnumPopup(UI.Orientation, _data.Orientation);
 			_data.AlphaThreshold = EditorGUILayout.Slider(UI.AlphaThreshold, _data.AlphaThreshold, 0f, 1f);
+
+			EditorGUILayout.Space();
 			_data.Monospaced = EditorGUILayout.Toggle(UI.Monospaced, _data.Monospaced);
+			_data.Ascent = UIUtils.FloatFieldMin(UI.Ascent, _data.Ascent, 0f);
+			_data.Descent = UIUtils.FloatFieldMin(UI.Descent, _data.Descent, 0f);
+
+			GUILayout.BeginHorizontal();
+			_data.LineSpacing = EditorGUILayout.FloatField(UI.LineSpacing, _data.LineSpacing);
+			if (GUILayout.Button(UI.GuessLineSpacingButton, Styles.MiniButton)) GuessLineSpacing();
+			GUILayout.EndHorizontal();
 
 			EditorGUILayout.Space();
 			DrawCharacterSetDropDown();
@@ -92,6 +104,7 @@ namespace dev.klebersilva.tools.bitmapfontcreator
 			DrawBottomMenu();
 
 			if (!string.IsNullOrEmpty(_error)) ShowCurrentError();
+			if (showPreview) TexturePreviewPopup.Open(_data, _prefs);
 		}
 
 		private void ShowCurrentError()
@@ -104,8 +117,15 @@ namespace dev.klebersilva.tools.bitmapfontcreator
 		{
 			EditorGUI.BeginChangeCheck();
 			_data.Texture = EditorGUILayout.ObjectField(UI.Texture, _data.Texture, typeof(Texture2D), false) as Texture2D;
-			if (EditorGUI.EndChangeCheck())
-				_guessCache = Vector2Int.zero;
+			if (EditorGUI.EndChangeCheck()) _guessRowsColsCache = Vector2Int.zero;
+		}
+
+		private bool DrawPreviewButton()
+		{
+			GUI.enabled = _data.Texture != null && _data.Rows > 0 && _data.Cols > 0;
+			var value = GUILayout.Button(UI.PreviewButton, Styles.MiniButton);
+			GUI.enabled = true;
+			return value;
 		}
 
 		private void DrawCharacterSetDropDown()
@@ -136,15 +156,10 @@ namespace dev.klebersilva.tools.bitmapfontcreator
 
 		private void DrawCreateFontButton()
 		{
-			GUILayout.BeginHorizontal();
-			GUILayout.FlexibleSpace();
 			GUI.color = Color.cyan;
-			if (GUILayout.Button(UI.CreateButton, Styles.CreateButton))
+			if (UIUtils.CenteredButton(UI.CreateButton, Styles.CreateButton))
 				BitmapFontCreator.TryCreateFont(_data, _prefsView.Model.WarnOnReplaceFont, out _error);
-
 			GUI.color = Color.white;
-			GUILayout.FlexibleSpace();
-			GUILayout.EndHorizontal();
 		}
 
 		private void DrawBottomMenu()
@@ -172,12 +187,28 @@ namespace dev.klebersilva.tools.bitmapfontcreator
 				return;
 			}
 
-			_guessCache = _guessCache.x == 0 || _guessCache.y == 0
+			_guessRowsColsCache = _guessRowsColsCache.x == 0 || _guessRowsColsCache.y == 0
 				? BitmapFontCreator.GuessRowsAndCols(_data.Texture)
-				: _guessCache;
+				: _guessRowsColsCache;
 
-			_data.Cols = _guessCache.y;
-			_data.Rows = _guessCache.x;
+			_data.Cols = _guessRowsColsCache.y;
+			_data.Rows = _guessRowsColsCache.x;
+		}
+
+		private void GuessLineSpacing()
+		{
+			if (_data.Texture == null)
+			{
+				Debug.LogWarning("Texture cannot be null");
+				return;
+			}
+			if (_data.Cols <= 0 || _data.Rows <= 0)
+			{
+				Debug.LogWarning("Rows and Cols must be greater than 0");
+				return;
+			}
+
+			_data.LineSpacing = BitmapFontCreator.GuessLineSpacing(_data.Texture, _data.Rows);
 		}
 	}
 }